requirements.txt
asgiref==3.8.1
Django==4.2.17
django-safedelete==1.4.0
djangorestframework==3.15.2
packaging==24.2
psycopg==3.1.8
psycopg-binary==3.1.8
sqlparse==0.5.2
typing_extensions==4.12.2
python-dateutil==2.9.0
django-cors-headers>=3.9.0
djangorestframework-simplejwt
gunicorn
